spirv: fix build for C++20
This codebase is using C++17, but to remain compatible with C++20,
I need to remove this.
Starting C++20, an aggregate is required to have no user-declared
constructors, vs before, no user-provided constructors.

So even if if the only the default constructor is provided, this
is enough to prevent us from using this class as an aggregate in c++20.

Signed-off-by: Nathan Gauër <brioche@google.com>