[CMAKE] Cleanup logic for enabling SPIR-V
This just cleans up the logic for enabling and disabling building SPIR-V
and the SPIR-V tests. It is _extremely_ unusable to ignore a
user-defined value that is specified as an option. In the current code
if a user enables SPIR-V tests but disables SPIR-V codegen, the
user-specified option to enable codegen gets ignored.