Fix bad arg indexing after array of struct in semantic assignment
Semantic index assignment for flattened arguments would incorrectly increment the argument index after an array of struct field was encountered. It would misinterpret any array as a basic type because it wasn't a struct or matrix type.

This fixes the bug by drilling into the array element before checking the type.