Allow unsigned int for nonzero base instance (#6120)
A regression was introduced by Refactor create stage vars (#6059)
because it created the BaseInstance variable as an int in all cases it
was created for the option vk-support-nonzero-base-instance enabled. In
some cases it was used as an unsigned int.

This comment get the type for BaseInstance from the type of the instance
id variable, and adds a test for unsigned.

Fixes #6118.