[SPIR-V] Remove dead code and correct Float16 docs
The docs incorrectly stated that the `SPV_AMD_gpu_shader_half_float`
extension was added when half types were used. This is now corrected to
show that the Float16 capability is added instead.

The remaining code that could have added that extension in certain cases
has been dead code since #2503 since none of the instructions checked in
that list are ones that are emitted by DXC, so it is removed to clean
up.

Follow up from https://github.com/KhronosGroup/SPIRV-Tools/pull/5519