[SPIRV] Implement out variables in patch control functions
The documentation for the patch control functions only mentions

> The outputs are usually defined by a structure and is
identified by HS_CONSTANT_DATA_OUTPUT in this example; the structure
depends on the domain type and would be different for triangle or
isoline domains.

This is why we did not realize that we needed to implement out variables
as well. This commit implement the alternate method of having an output.

Fixes #3743