Require outputtopology attribute on Hull shader
There was an existing check in DXIL validation that failed when the
outputtopology was not set for a Hull shader. Rather than duplicate this
check in the SPIR-V backend/validator, it seems correct to check that it
is present in Sema.

This PR needs review from someone who is more familiar with the HLSL spec
for hull shaders to verify that this check is correct though. I can't
say for sure from the HLSL docs.

Fixes #3740