[SPIRV] Fix the code generated for dot4add_u8packed and dot4add_i8packed (#6097)
This is a follow-up for
https://github.com/microsoft/DirectXShaderCompiler/pull/5959

The SPIR-V spec
[requires](https://registry.khronos.org/SPIR-V/specs/unified1/SPIRV.html#OpSDot)
that `OpSDot` and `OpUDot` have an extra Packed Vector Format operand
when Vector 1 and Vector 2 are scalar integer types. My previous
implementation didn't add that operand - sorry about that. The generated
code worked on NVIDIA GPUs but produced incorrect ouptut on AMD and
resulted in a crash on Intel ARC. This PR fixes the generated code so
that it works on all three vendors' GPUs.