Remove SPIRV test case from Sema tests related to DS/HS patch size (#6166)
The tests are for an error message that originates from Sema HLSL, it is
a not a spirv-specific code. The `-spirv` test case does not bring any
additional value. And since these tests are not marked spirv-only (as
they should not be), they are failing when the compiler is built without
SPIRV support.