Fix empty control statement warnings when asserts are diabled on MSVC (#6197)
When asserts are disabled (LLVM_ENABLE_ASSERTIONS=0), DXASSERT* macros
are defined to nothing, resulting in no statements followed by a
semicolon. This results in:

warning C4390: ';': empty controlled statement found; is this the
intent?

Which fails because warnings are treated as errors.

Fix by adding the standard "do {} while(0)". Also opportunistically
remove the extra semicolon for non-WIN32 builds.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>