[SPIR-V] Fix literal float assertion
Fix #6230

The literal type visitor is asserting that the AST result type of an
OpLoad should not be literal. The following expression was
encountering this assertion.

float_vec_4 *= condition ? -1.0 : 1.0;

This patch visits OpVectorTimesScalar operations to change the
AST result type of an instruction from a literal type to a non-literal
type where possible.

Also, fix some mispellings I encountered while navigating the code.