Add testing infrastructure to test SPIR-V codegen from lib call (#6152)
We recently removed the FileTest infrstructure because we were able to
replace all of the tests with LIT tests. However, we now have to test
the a case were the source code is in memory and not in a file (see
`FileTest`, buti specializes it so that it cannot be given a file name,
and
remove the uses of EFFCEE.

One test is added as an example on how to use a REGEX matcher to check
the assemble output.

Other options we considers:

1. Adding new infrastructure based on the code in HLSL/CompilerTest.cpp.
   However that was more complex that we need. I wanted a simple
interface so that it is easy for people to add new tests. I also wanted
   to make it spir-v specific.
2. Modifying the DXC executable to not pass along the file name. This
   was rejected because it would place an extra option, just for
   testing, the public facing executable.
3. Have the dxc executable accept `-` at the input file to have it read
   the input file from stdin. Similar to 2. Rejected for similar
   reasons.