Add OptFeatureInfo_UsesDerivatives, fix some DxilShaderFlags
Adding OptFeatureInfo_UsesDerivatives to track per-function derivative usage,
otherwise we can't set DerivativesInMeshAndAmpShaders or min shader target properly.

This uses a feature bit in a special optional set of flags defined by the runtime to be ignored if set and unrecognized.
Since it's not a real feature requirement but a special flag used in computing feature requirements after accounting for all functions at link time, this is a good place for this special flag.