Move ShaderCompatInfo to DxilModule, recurse calls
ShaderCompatInfo was used to track per-function feature flag and min target for RDAT.
But internal calls were not traversed and the global feature flag collection had to do all the same work as the RDAT writer did redundantly.

This change moves this into DxilModule, where it can be computed once and cached for both the CollectShaderFlagsForModule, as well as the RDAT per-function lookup.
Along the way, DxilModule& needed to be made non-const for DxilRDATWriter so it could update the compat info, in case it was unwritten.