[SPIRV] Reimplement flat conversion. (#6189)
The implementation of flat conversion in the spir-v codegen is very
adhoc. This changes the implemetation to match what is done in DXIL. For
now I have only reimplemented to catch all case. I will leave the
special cases as they seem to be correct for what they do, and will
generate less verbose code.

Fixes #4906

---------

Co-authored-by: Nathan Gauër <github@keenuts.net>