Resolve PRERELEASE-TODO cases related to deep RDAT validation
These TODOs were intended for if we were to implement deep RDAT validation rather than relying on the exact binary comparison.

This isn't happening for SM 6.8, so these are no longer PRERELEASE-TODOs.

The comment in RDAT_Macros.inl didn't really belong there in the first place, so I removed it.