Remove early skip of unused GV causing worklist ordering differences
Early skip of unused GV caused worklist ordering differences, which could result in differently ordered allocas, which is causing optimization regressions in backend.

Reverting this part of prior change (f80b8d6) from PR #5259,
since it was not required for the original purpose of removing dead GVs from the staticGVs set.

Unfortunately, it is proving prohibitively difficult to craft a minimal repro for a test case that demonstrates the difference.