[SPIR-V] Emit error for 16-bit texture types
According to the Vulkan spec:
OpTypeImage must declare a scalar 32-bit float, 64-bit integer, or 32-bit integer type for the “Sampled Type” (RelaxedPrecision can be applied to a sampling instruction and to the variable holding the result of a sampling instruction)

We want to avoid translating a true HLSL 16-bit floating point texture
object to a 32-bit texture object with relaxed precision, as this will
break backwards compatibility if Vulkan does support 16-bit texture
types in the future.

This change adds an error message in cases where the code generated
would otherwise fail validation for the above reasons.

Closes #4372