[SPIRV] Fix out of bounds vector access (#6253)
One of the current test has an out of bounds memory access in a vector.
We access `scalars[0]` vector after all of the elements have been
removed. This is used to get the source location. The solution is to get
the source location earlier, and use the local variable.

No new test is needed because a current test exposed the problem on a
platform and build that asserts when the access happens.