Address feedback; Improve testing
One notable change is the elimination of the optimization in ComputeShaderCompatInfo where a function would be skipped if it already had ShaderCompatInfo.
This was incorrect, since this is happening inside the same loop that iterates intrinsics, which could add the compat info based on intrinsic usage.
That doesn't mean flags have been computed for the function, so I removed this and clear the map at the beginning of ComputeShaderCompatInfo to ensure it's fresh.

I also added IR tests, and ones that reorder intrinsics and the called function to ensure function ordering doesn't cause a problem.

I made these and the sm66_derivatives.hlsl test they were based on go back to SM 6.6, since there was another detail on that validator version that was different - it didn't set min shader model properly for lib functions.
The 1.6 validator is the one included in the latest Windows SDK, so going back this far seems good.

I also added the OptFeatureInfo flags to disassembly.