Warn on ignored attributes on library non-entry points (#6243)
For every attribute that is only meaningful on an entry-point, produce a
warning indicating that this could be a problem.
The attributes are ignored, so it's harmless, but it still might
indicate a mistake on the part of the user.
Previously, some of these would cause the function to be implicitly made
a certain kind of entry point.
These warnings will soften the impact of that change.

Fixes #6100

---------

Co-authored-by: Joshua Batista <jbatista@microsoft.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
(cherry picked from commit 53b2930a1b3ad79cbc060d5db451906a30c2691b)