Add CODEOWNERS to release (#6257)
As per the schedule in #6203, we will begin AskMode on 2/12, at which
time all PRs will be required to have at least one sign off from a
member of @microsoft/hlsl-release.

Adding this CODEOWNERS file allows us to have GitHub enforce that
requirement.

Note: This PR does not enable AskMode, it just puts the infrastructure
in place. AskMode will not be enabled until 2/12.