[SPIRV] Using 64-bit int literals in flat conversion (#6283)
During flat convertion, we have to pick the bit width for literals
because they could be used in multiple places as different sizes, and
that breaks the literal visitor.  We currently use 32-bit types because
that is the most common use case, and we don't want to unnecessarily
add a capability.

However, if we pick a 64-bit width, we should be able to fold all of
the operations on literals. If they are only used as 32-bit value, then
ultimatlly fold them to 32-bit constants and remove the 64-bit
capabilities.

This commit updates the spir-v submodule to pull in the 64-bit interger
folding changes, and then starts to use 64-bit integer literal in flat
conversion.

Note that the problem still exists for floating point values. Spirv-opt
does not fold OpFConvert yet, we still treat the literals like a float.

Fixes #6188

---------

Co-authored-by: Natalie Chouinard <chouinard.nm@gmail.com>