[Sema] Only emit warnings on entry point attrs with missing shader attr for libraries (#6284)
In implementing this PR:
https://github.com/microsoft/DirectXShaderCompiler/pull/6243
somehow, the condition to only emit the warning for library shaders was
left out.
This PR is a follow up on #6243 to properly restrict the emission of
this warning for non-library shaders.
This should properly fix #6100

(cherry picked from commit dbf60d777156b98ea2f8cafc8336bc62bf65213c)