[SPIRV] Cast SV_CULLPRIMITIVE to uint
When SV_CULLPRIMITIVE is the output of a mesh shader, its type is
changed to uint because a Vulkan interface variable cannot be a bool.
However, we do not change the access to the variable to match.

This change will modify stores to the variable to be stores of a uint
instead of a bool.

Fixes #4602