[Sema] Allow int arguments for attributes to be enum values
The SPIR-V backend uses `ValidateAttributeIntArg` for the
`vk::ext_builtin_input` and `vk::ext_builtin_output` attrs, as well as
some others. It is convenient for users to be able to pass in an enum
value to this attribute, for example:

```
enum class BuiltIn {
  HelperInvocation = 23
};

[[vk::ext_builtin_input(BuiltIn::HelperInvocation)]]
static const bool gl_HelperInvocation;
```

If you think `ValidateAttributeIntArg` *should* warn if a user passes in
an enum value, I'll create a separate function for validating
`vk::ext_builtin_input` instead.

Fixes #6337