PIX: Cp of "Prevent ray query handle writes to PIX debug UAV" (#6362)
Original PR:
https://github.com/microsoft/DirectXShaderCompiler/pull/6309

Original commit:

https://github.com/microsoft/DirectXShaderCompiler/commit/5a3178590581b4a7deea17dd45b6c4f1b51edb53

AllocateRayQuery returns an int, but it's not really an int: it's a
handle to the query. Thus, it's not sensible for the PIX debugging
instrumentation to attempt to write it out to the debug-data-UAV. The
previous code did an explicit check against the type of the value to be
written, but if that value is actually a phi itself, then that check
would fail. So now we recursively run through the phi values looking to
see if any of its antecedents are a phi, and if so, refuse to send its
value to the UAV.

(cherry picked from commit 5a3178590581b4a7deea17dd45b6c4f1b51edb53)