sema: check input patch size mismatch earlier.
The check was done in the DXIL backend, and not in the SPIR-V backend.
This causes the issue to be ignored on the SPIR-V side.
Moved the check earlier to SEMA so a single check works for both
backends.

I had to split the existing HLSL check as it doesn't support having
2 run lines (AFAIK).

Fixes #3739

Signed-off-by: Nathan Gauër <brioche@google.com>