Fixed invalid debug metadata for LexicalBlock in include files. (#6373)
No longer normalizing paths that are baked into `DIFile` (as well as the
debug module file list and PDB source info).

The previous path normalization change introduced a bug in debug info
where instead of generating a normal `DILexicalBlock`, we generated a
`DILexicalBlockFile`. This was because the file paths written into the
`DIFile`'s were normalized, but the file list in `SourceManager` was
left unchanged. `CGDebugInfo::setLocation` checks whether the `DIFile`
path matches the current file, and due to the normalization differences,
assume the file has changes and compensates by generating a new
`DILexicalBlockFile` on the scope stack.

An alternative to this PR would be to change `CGDebugInfo::setLocation`
to normalize the `SourceManager` path before comparing; however, there
may be other unknown code paths that could have similar issues. Another
alternative is to normalize the path list in `SourceManager`; however,
this has other ramifications and require many more drastic changes.

By no longer normalizing the paths written into `DIFile`, we revert to
the behaviour where the paths in `DIFile` come directly from
`SourceManager`, so there is no chance of a mismatch. Furthermore, by
avoiding normalizing any paths that are written to PDBs, we do not bake
any potential normalization bugs into any test content.