[SPIR-V] Disallow variable offsets in texture Load
The Load function's optional Offset parameter must be an immediate
value to match the behavior of the DXIL backend. The existing logic in
the SPIR-V backend to evaluate it as a variable offset if necessary
generated invalid SPIR-V that violated the
`VUID-StandaloneSpirv-Offset-04663` rule:

> Image operand Offset must only be used with OpImage*Gather instructions

The logic for processing the argument as a variable value has been
removed and replaced with an error message that matches DXIL's.

Note that `handleOffsetInMethodCall` is left unmodified because it is
also used to process HLSL `.Gather()` calls, which are lowered to
`OpImageGather` SPIR-V instructions, which can take non-const Offsets as
image operands.

Fixes #6149