AddTypeAnnotation: skip incomplete type
An incomplete type indicates an unused type, so it should not require a type annotation.

The incomplete type case is hit due to new code annotating object method return types.  If the return type is an incomplete type, the method must never be called, the type is never completed, and the type annotation cannot be created.