[SPIR-V] Block semantic annotation reuse for input
The MSDN spec is not very clear regarding input parameter
aliasing, BUT DXIL & MS agrees (See #3737) using the same
semantic annotation twice should be disallowed.

DXIL currently disallows it for most, and due to a bug, allows
some compute related semantics to be aliased.
SPIR-V did allowed aliasing, but it was a implemented as a hack causing
typing issues if the type changed between the 2 parameters using the
same semantic annotation.

To align more closely with DXIL, and the "spec", we are not disallowing
all semantic attribute reuse for input parameters.

Fixes #3737

Signed-off-by: Nathan Gauër <brioche@google.com>