Allow const vars in attribute parameters
HLSL does not allow const vars to be evaluated in integer constant
expressions - this change relaxes this restriction in the case of
attribute integer parameter values. Specifically, a const var with
integer constant expression initializer is accepted, and the value
of the initializer is used as the value of the attribute parameter.
This (expermental) behavior may be useful in cases where Vulkan
specialization constants are used as an attribute parameter value.