Remove assert on illegal DXIL overload
Certain HL ops with no double overload will generate double overloads for literal types.
These are lowered to double DXIL overloads, some of which are not legal for these ops.
We currently rely on illegal intermediate DXIL op overloads for constant evaluation.

Disabling this assert allows test cases to be written to verify the current behavior for these operations.