Remove unnecessary and sensitive check.
What matters is 'error: DXIL intrinsic overload must be valid.'
It's unnecessary to check the specific intrinsic reported,
partly because it will only report the first one in function declaration order,
which could change with changes to other code.

The purpose of the test is to make sure the validator gracefully caught the invalid DXIL overload.