Add tests for literal overloads of fmod, length, and normalize (#6437)
Certain HL ops with no double overload will generate double overloads
for literal types. These are lowered to double DXIL overloads, some of
which are not legal for these ops. We currently rely on illegal
intermediate DXIL op overloads for constant evaluation. If constant
evaluation fails, we end up with illegal DXIL overloads in final DXIL,
which is caught by the validator.

The prior revert restored the ability to run these scenarios (with
asserts disabled).

This change adds tests for #6419 so automated testing will test across
configurations and prevent regressions of this scenario in this branch.